==> code/02_modeling.R <==
library(mgcv)
library(dplyr)
library(gratia)
library(ggplot2)

all_combos <- expand.grid(data = c("Alaska", "WC")[1],
                          scale = c("port"),
                          split_wc = c("north_south","one_area")[1],
                          cs_sensitivity = c(TRUE,FALSE),
                          include_all_vessels = c(TRUE,FALSE))

for(ii in 1:nrow(all_combos)) {
  data <- all_combos$data[ii]
  scale <- all_combos$scale[ii]
  #Split west coast into north/south of 40 10?
  split_wc <- all_combos$split_wc[ii]
  cs_sensitivity <- all_combos$cs_sensitivity[ii]
  include_all_vessels <- all_combos$include_all_vessels[ii]
  #label that appends file names with whether we subset only to vessels present both before and after CS ("stayers")
  if(include_all_vessels)
  {
    vessel_label = "allvessels"
  }else{
    vessel_label = "singleport"
  }

  if(cs_sensitivity)
  {
    cs_sens_label = "stayers"
  }else{
    cs_sens_label = "allvessels"
  }

  # these dataframes get created by 01_regional_summaries.r
  if(include_all_vessels==TRUE) {
    area_permit_cog = readRDS(paste0("data/",data,"_",cs_sens_label,"_",scale,"_cog",".rds"))
    area_permit_cog_ind = readRDS(paste0("data/",data,"_",cs_sens_label,"_",scale,"_cog-ind",".rds"))
    eff_days = readRDS(paste0("data/",data,"_",cs_sens_label,"_",scale,"_days",".rds"))
    eff_days_ind = readRDS(paste0("data/",data,"_",cs_sens_label,"_",scale,"_days-ind",".rds"))
    model_runs <- c("inertia","inertia-ind",
      "eff_days","eff_days-ind",
      "lat","lat-ind",
      "lon","lon-ind")
  } else {
    haul_dist = readRDS(paste0("data/",data,"_",cs_sens_label,"_",scale,"_hauldist",".rds"))
    haul_dist_ind = readRDS(paste0("data/",data,"_",cs_sens_label,"_",scale,"_hauldist-ind",".rds"))
    model_runs <- c("haul_dist","haul_dist-ind")
  }

  # using area permit_cog as an example, the predictor variables here are:
  # year, port (v) / subarea, and sector2.

  # Example questions:
  # - are trends in inertia explained better by region (subarea), sector,
  # or is there no common pattern?
  plot_list = list()
  pred_list = list()

  for(run in model_runs) {

    if(run %in% c("inertia","lat","lon")) dat = area_permit_cog
    if(run %in% c("inertia-ind","lat-ind","lon-ind")) dat = area_permit_cog_ind
    if(run=="haul_dist") dat = haul_dist
    if(run=="haul_dist-ind") dat = haul_dist_ind
    if(run=="eff_days") dat = eff_days
    if(run=="eff_days-ind") dat = eff_days_ind

    if(run%in% c("inertia","inertia-ind")) {
      ylabel = "ln sqrt(inertia)"
      dat$response = log(dat$total_sd)
    }
    if(run%in% c("lat","lat-ind")) {
      ylabel = "latitude"
      dat$response = dat$lat_mean
    }
    if(run%in% c("lon","lon-ind")) {
      ylabel = "longitude"
      dat$response = dat$lon_mean
    }
    if(run%in% c("haul_dist","haul_dist-ind")) {
      ylabel = "ln distance"
      dat$response=log(dat$mean_haul_dist)
    }
    if(run%in% c("eff_days","eff_days-ind")) {
      ylabel = "ln eff. days"
      dat$response=log(dat$effective_days)
    }

    dat = dplyr::filter(dat, !is.na(response), is.finite(response)) %>%
      dplyr::filter(!is.na(sector2), !is.na(subarea))

    dat$port = as.factor(dat$v)
    # only use ports with 10 or more years of data
    dat = group_by(dat, port) %>%
      dplyr::mutate(nyear = length(unique(year))) %>%
      dplyr::filter(nyear>=10) %>%
      dplyr::select(-nyear)

    # TODO: discuss appropriate weights. What level of variability / sample size do we care about?
    dat$weights = 1
    #if(run %in% c("area_permit_cog-ind", "haul_dist-ind","eff_days-ind")) dat$weights = dat$n
    #dat = dplyr::filter(dat, total_cv > 0, !is.na(total_cv)) %>%
    #  dplyr::mutate(tot_weights = 1/total_cv,
    #                weights = 1)#tot_weights/mean(tot_weights))

    # filter out combinations that aren't common in the data
    dat$sector2 = as.factor(dat$sector2)
    dat$subarea = as.factor(dat$subarea)
    dat$sector_subarea = as.factor(paste(dat$sector2, dat$subarea))
    not_rare = dplyr::group_by(dat, sector_subarea) %>%
      dplyr::summarise(n = n()) %>%
      dplyr::filter(n >= 10) # 40 = port-year combinations
    #Need to adjust this for WC so that we don't exclude at-sea hake, which only has one "port"
    if(data == "WC")
    {
      #dat = dplyr::filter(dat, sector_subarea %in% not_rare$sector_subarea | sector2 == "At-sea hake")
      NULL #Filtering only to common combinations in the WC data removes LE sablefish south for the distance from port calculations, and I don't think we want that
    }else{
      dat = dplyr::filter(dat, sector_subarea %in% not_rare$sector_subarea)
    }
    dat$sector_subarea = as.factor(as.character(dat$sector_subarea))

    # binary indicator for rationalization
    dat$catch_share = 0 # 0 before, 1 after
    if(data == "Alaska") {
      # longline misc. groundfish    pelagic trawl         rockfish
      dat$catch_share[which(dat$sector2 == "rockfish" & dat$year >= 2005)] = 1
      dat$catch_share[which(dat$sector2 == "longline" & dat$year >= 1995)] = 2
    }else{
      dat$catch_share[which(dat$sector2 == "LE/CS Trawl" & dat$year >= 2011)] = 1
      dat$catch_share[which(dat$sector2 == "At-sea hake CP" & dat$year >= 2011)] = 2
      dat$catch_share[which(dat$sector2 == "At-sea hake MS" & dat$year >= 2011)] = 3

    }
    dat$catch_share = as.factor(dat$catch_share)

    # global year effect, with port level smooths (port = v)
    fit = list()
    K = 10 # knots for year effects, for factor smooths

    # New null model, response varies by subarea
    n_unique_subarea <- length(unique(dat$subarea))
    n_unique_sector <- length(unique(dat$sector2))
    fit[[1]] = gam(response ~ catch_share + s(year,bs="ps"), weights = weights,data = dat)
    if(n_unique_subarea > 1 & n_unique_sector > 1) {
      # statement needed because some AK data only has 1 area
      fit[[2]] = gam(response ~ catch_share + sector2 + subarea + s(year,bs="ps"), weights = weights,data = dat)
      fit[[3]] = gam(response ~ catch_share + sector2 + subarea + s(year, subarea, k=K, bs="fs", m=2), weights = weights,data = dat)
      fit[[4]] = gam(response ~ catch_share + sector2 + subarea + s(year, sector2, k=K, bs="fs", m=2), weights = weights,data = dat)
      fit[[5]] = gam(response ~ catch_share + sector2 + s(year, subarea, k=K, bs="fs", m=2) + s(year, sector2, k=K, bs="fs", m=2), weights = weights,data = dat)
      fit[[6]] = gam(response ~ catch_share + sector2 + subarea + s(year, sector_subarea, k=K, bs="fs", m=2), weights = weights,data = dat)
    }
    # save models for later summarizing
    saveRDS(fit, paste0("output/gams_",scale, "_",run,"_",data,"_",cs_sens_label,"_", split_wc,".rds"))

    # Add random intercepts to ports
    # fit[[7]] = gam(response ~ s(year,bs="ps") + s(port,bs="re"), weights = weights,data = dat)
    # fit[[8]] = gam(response ~ sector2 + subarea + s(year,bs="ps") + s(port,bs="re"), weights = weights,data = dat)
    # fit[[9]] = gam(response ~ sector2 + s(year, subarea, k=K, bs="fs", m=2) + s(port,bs="re"), weights = weights,data = dat)
    # fit[[10]] = gam(response ~ subarea + s(year, sector2, k=K, bs="fs", m=2) + s(port,bs="re"), weights = weights,data = dat)
    # fit[[11]] = gam(response ~ s(year, subarea, k=K, bs="fs", m=2) + s(year, sector2, k=K, bs="fs", m=2) + s(port,bs="re"), weights = weights,data = dat)
    # fit[[12]] = gam(response ~ s(year, sector_subarea, k=K, bs="fs", m=2) + s(port,bs="re"), weights = weights,data = dat)

    # create new data to predict on -- representing mean for each area-sector
    newdata = expand.grid(year = unique(dat$year),
                          sector2 = unique(dat$sector2),
                          port = unique(dat$port),
                          catch_share = 0)
    if(data == "Alaska") {
      # longline misc. groundfish    pelagic trawl         rockfish
      newdata$catch_share[which(newdata$sector2 == "rockfish" & newdata$year >= 2005)] = 1
      newdata$catch_share[which(newdata$sector2 == "longline" & newdata$year >= 1995)] = 2
    }else{
      newdata$catch_share[which(newdata$sector2 == "LE/CS Trawl" & newdata$year >= 2011)] = 1
      dat$catch_share[which(dat$sector2 == "At-sea hake CP" & dat$year >= 2011)] = 2
      dat$catch_share[which(dat$sector2 == "At-sea hake MS" & dat$year >= 2011)] = 3
      }
    newdata$catch_share = as.factor(newdata$catch_share)
    # add block

    newdata = dplyr::filter(newdata,!is.na(sector2), !is.na(port))
    port_area = dplyr::group_by(dat, port) %>% dplyr::summarize(subarea = subarea[1])
    newdata = dplyr::left_join(newdata, port_area)
    newdata$sector_subarea = as.factor(paste(newdata$sector2, newdata$subarea))
    newdata = dplyr::filter(newdata,sector_subarea %in% unique(dat$sector_subarea))

    # This block is just making predictions for each model and binding all together
    model_pred = predict(fit[[1]], newdata = newdata, se.fit=TRUE)
    newdf = cbind(newdata, model_pred)
    newdf$model = 1
    newdf$run = run
    newdf$scale = scale
    newdf$region = data
    all_pred = newdf
    if(n_unique_subarea > 1 & n_unique_sector > 1) {
      for(i in 2:6) {
        model_pred = predict(fit[[i]], newdata = newdata, se.fit=TRUE)
        newdf = cbind(newdata, model_pred)
        newdf$model = i
        newdf$run = run
        newdf$scale = scale
        newdf$region = data
        all_pred = rbind(all_pred, newdf)
      }
    }
    saveRDS(all_pred, paste0("output/predictions_",scale, "_",run,"_",data,"_",cs_sens_label,"_", split_wc,".rds"))

    # make initial plots using model with lowest AIC
    model_pred <- predict(fit[[which.min(lapply(fit,AIC))]], newdata=newdata, se.fit=TRUE) # predict to new port -- will be same as mean
    newdata = cbind(newdata, model_pred)

    newdata = dplyr::group_by(newdata, sector_subarea, year) %>%
      dplyr::summarize(
        sector2 = sector2[1],
        subarea = subarea[1],
        n = n(), # n is number of ports
                       fit = mean(fit,na.rm=T),
                       se = sqrt((1/(n*n)) * sum(se.fit^2))
                       )
    #keep track of which model had lowest AIC
    best_model <- which.min(lapply(fit,AIC))

    # look at individual smooths
    #sm = get_smooth(fit[[which.min(lapply(fit,AIC))]], "year")
    #sm = smooth_estimates(fit[[5]], data=newdata)
    #sm = dplyr::filter(sm, smooth=="s(year,sector_subarea)")
    #model_pred <- predict(fit[[which.min(lapply(fit,AIC))]], newdata=newdata, se.fit=TRUE) # predict to new port -- will be same as mean

    # save plots
    if(run==model_runs[[1]]) {
      counter = 1
    } else {
      counter = counter + 1
    }
    plot_list[[counter]] = ggplot(newdata, aes(year, fit, col=sector2, fill=sector2)) +
      geom_line() +
      geom_ribbon(aes(ymin=fit-2*se, ymax=fit+2*se),alpha=0.5) +
      facet_wrap(subarea~sector2,scale="free_y") +
      theme_bw() + xlab("") + ylab(ylabel) + ggtitle(paste0(run, " (model ", best_model, ")"))

    pred_list[[counter]] = newdata

  } # end loop over runs

  ## Make summary table of AIC stats for paper
  model_results = NULL
  for(run in model_runs) {
    fit = readRDS(paste0("output/gams_",scale, "_",run,"_",data,"_",cs_sens_label,"_", split_wc,".rds"))
    # extract AIC etc
    df = data.frame("Model"=1:6, "Run" = run, "AIC" = unlist(lapply(fit,AIC)))
    df$Scale = ifelse(length(grep("ind",run))==0,"Aggregate","Individual")
    df$AIC = df$AIC - min(df$AIC)
    # bind all results together
    if(is.null(model_results)) {
      model_results = df
    } else {
      model_results = rbind(model_results, df)
    }
  }
  model_results$Metric = "Season"
  model_results$Metric[which(model_results$Run %in% c("inertia","inertia-ind"))] = "Inertia"
  model_results$Metric[which(model_results$Run %in% c("haul_dist","haul_dist-ind"))] = "Distance"
  model_results$Metric[which(model_results$Run %in% c("lat","lat-ind"))] = "Latitude"
  model_results$Metric[which(model_results$Run %in% c("lon","lon-ind"))] = "Longitude"
  model_results$Area = data
  saveRDS(model_results, paste0("output/aic_",scale,"_",data,"_",cs_sens_label,"_", vessel_label,"_",split_wc,".rds"))

  pdf(paste0("output/",scale, "_gams_",data,"_",cs_sens_label,"_", vessel_label,"_",split_wc,".pdf"))
  for(i in 1:length(plot_list)) plot_list[i]
  dev.off()


}
